refactor(protocol): isolate legacy item fanout
