+++ codex-rs/protocol/src/items.rs
@@ -10,26 +10,12 @@ use crate::models::ResponseItem;
 use crate::models::WebSearchAction;
 use crate::openai_models::ReasoningEffort as ReasoningEffortConfig;
 use crate::parse_command::ParsedCommand;
-use crate::protocol::AgentMessageEvent;
-use crate::protocol::AgentReasoningEvent;
-use crate::protocol::AgentReasoningRawContentEvent;
 use crate::protocol::AgentStatus;
 use crate::protocol::CollabAgentRef;
-use crate::protocol::ContextCompactedEvent;
-use crate::protocol::EventMsg;
 use crate::protocol::ExecCommandSource;
 use crate::protocol::FileChange;
-use crate::protocol::ImageGenerationEndEvent;
-use crate::protocol::McpInvocation;
-use crate::protocol::McpToolCallBeginEvent;
-use crate::protocol::McpToolCallEndEvent;
-use crate::protocol::PatchApplyBeginEvent;
-use crate::protocol::PatchApplyEndEvent;
 use crate::protocol::PatchApplyStatus;
 use crate::protocol::SubAgentActivityKind;
-use crate::protocol::UserMessageEvent;
-use crate::protocol::ViewImageToolCallEvent;
-use crate::protocol::WebSearchEndEvent;
 use crate::user_input::ByteRange;
 use crate::user_input::TextElement;
 use crate::user_input::UserInput;
@@ -387,10 +373,6 @@ impl ContextCompactionItem {
             id: uuid::Uuid::new_v4().to_string(),
         }
     }
-
-    pub fn as_legacy_event(&self) -> EventMsg {
-        EventMsg::ContextCompacted(ContextCompactedEvent {})
-    }
 }
 
 impl Default for ContextCompactionItem {
@@ -406,20 +388,6 @@ impl UserMessageItem {
             client_id: None,
             content: content.to_vec(),
         }
-    }
-
-    pub fn as_legacy_event(&self) -> EventMsg {
-        // Legacy user-message events flatten only text inputs into `message` and
-        // rebase text element ranges onto that concatenated text.
-        EventMsg::UserMessage(UserMessageEvent {
-            client_id: self.client_id.clone(),
-            message: self.message(),
-            images: Some(self.image_urls()),
-            image_details: self.image_details(),
-            local_images: self.local_image_paths(),
-            local_image_details: self.local_image_details(),
-            text_elements: self.text_elements(),
-        })
     }
 
     pub fn message(&self) -> String {
@@ -609,134 +577,6 @@ impl AgentMessageItem {
             memory_citation: None,
         }
     }
-
-    pub fn as_legacy_events(&self) -> Vec<EventMsg> {
-        self.content
-            .iter()
-            .map(|c| match c {
-                AgentMessageContent::Text { text } => EventMsg::AgentMessage(AgentMessageEvent {
-                    message: text.clone(),
-                    phase: self.phase.clone(),
-                    memory_citation: self.memory_citation.clone(),
-                }),
-            })
-            .collect()
-    }
-}
-
-impl ReasoningItem {
-    pub fn as_legacy_events(&self, show_raw_agent_reasoning: bool) -> Vec<EventMsg> {
-        let mut events = Vec::new();
-        for summary in &self.summary_text {
-            events.push(EventMsg::AgentReasoning(AgentReasoningEvent {
-                text: summary.clone(),
-            }));
-        }
-
-        if show_raw_agent_reasoning {
-            for entry in &self.raw_content {
-                events.push(EventMsg::AgentReasoningRawContent(
-                    AgentReasoningRawContentEvent {
-                        text: entry.clone(),
-                    },
-                ));
-            }
-        }
-
-        events
-    }
-}
-
-impl WebSearchItem {
-    pub fn as_legacy_event(&self) -> EventMsg {
-        EventMsg::WebSearchEnd(WebSearchEndEvent {
-            call_id: self.id.clone(),
-            query: self.query.clone(),
-            action: self.action.clone(),
-        })
-    }
-}
-
-impl ImageGenerationItem {
-    pub fn as_legacy_event(&self) -> EventMsg {
-        EventMsg::ImageGenerationEnd(ImageGenerationEndEvent {
-            call_id: self.id.clone(),
-            status: self.status.clone(),
-            revised_prompt: self.revised_prompt.clone(),
-            result: self.result.clone(),
-            saved_path: self.saved_path.clone(),
-        })
-    }
-}
-
-impl FileChangeItem {
-    pub fn as_legacy_begin_event(&self, turn_id: String) -> EventMsg {
-        EventMsg::PatchApplyBegin(PatchApplyBeginEvent {
-            call_id: self.id.clone(),
-            turn_id,
-            auto_approved: self.auto_approved.unwrap_or(false),
-            changes: self.changes.clone(),
-        })
-    }
-
-    pub fn as_legacy_end_event(&self, turn_id: String) -> Option<EventMsg> {
-        let status = self.status.clone()?;
-        Some(EventMsg::PatchApplyEnd(PatchApplyEndEvent {
-            call_id: self.id.clone(),
-            turn_id,
-            stdout: self.stdout.clone().unwrap_or_default(),
-            stderr: self.stderr.clone().unwrap_or_default(),
-            success: status == PatchApplyStatus::Completed,
-            changes: self.changes.clone(),
-            status,
-        }))
-    }
-}
-
-impl McpToolCallItem {
-    pub fn as_legacy_begin_event(&self) -> EventMsg {
-        EventMsg::McpToolCallBegin(McpToolCallBeginEvent {
-            call_id: self.id.clone(),
-            invocation: McpInvocation {
-                server: self.server.clone(),
-                tool: self.tool.clone(),
-                arguments: (!self.arguments.is_null()).then(|| self.arguments.clone()),
-            },
-            connector_id: self.connector_id.clone(),
-            mcp_app_resource_uri: self.mcp_app_resource_uri.clone(),
-            link_id: self.link_id.clone(),
-            app_name: self.app_name.clone(),
-            template_id: self.template_id.clone(),
-            action_name: self.action_name.clone(),
-            plugin_id: self.plugin_id.clone(),
-        })
-    }
-
-    pub fn as_legacy_end_event(&self) -> Option<EventMsg> {
-        let result = match (&self.result, &self.error) {
-            (Some(result), _) => Ok(result.clone()),
-            (None, Some(error)) => Err(error.message.clone()),
-            (None, None) => return None,
-        };
-
-        Some(EventMsg::McpToolCallEnd(McpToolCallEndEvent {
-            call_id: self.id.clone(),
-            invocation: McpInvocation {
-                server: self.server.clone(),
-                tool: self.tool.clone(),
-                arguments: (!self.arguments.is_null()).then(|| self.arguments.clone()),
-            },
-            mcp_app_resource_uri: self.mcp_app_resource_uri.clone(),
-            connector_id: self.connector_id.clone(),
-            link_id: self.link_id.clone(),
-            app_name: self.app_name.clone(),
-            template_id: self.template_id.clone(),
-            action_name: self.action_name.clone(),
-            plugin_id: self.plugin_id.clone(),
-            duration: self.duration?,
-            result,
-        }))
-    }
 }
 
 impl TurnItem {
@@ -758,35 +598,6 @@ impl TurnItem {
             TurnItem::FileChange(item) => item.id.clone(),
             TurnItem::McpToolCall(item) => item.id.clone(),
             TurnItem::ContextCompaction(item) => item.id.clone(),
-        }
-    }
-
-    pub fn as_legacy_events(&self, show_raw_agent_reasoning: bool) -> Vec<EventMsg> {
-        match self {
-            TurnItem::UserMessage(item) => vec![item.as_legacy_event()],
-            TurnItem::HookPrompt(_) => Vec::new(),
-            TurnItem::AgentMessage(item) => item.as_legacy_events(),
-            TurnItem::Plan(_) => Vec::new(),
-            TurnItem::CommandExecution(_)
-            | TurnItem::DynamicToolCall(_)
-            | TurnItem::CollabAgentToolCall(_) => Vec::new(),
-            TurnItem::SubAgentActivity(_) => Vec::new(),
-            TurnItem::WebSearch(item) => vec![item.as_legacy_event()],
-            TurnItem::ImageView(item) => {
-                vec![EventMsg::ViewImageToolCall(ViewImageToolCallEvent {
-                    call_id: item.id.clone(),
-                    path: item.path.clone(),
-                })]
-            }
-            TurnItem::Sleep(_) => Vec::new(),
-            TurnItem::ImageGeneration(item) => vec![item.as_legacy_event()],
-            TurnItem::FileChange(item) => item
-                .as_legacy_end_event(String::new())
-                .into_iter()
-                .collect(),
-            TurnItem::McpToolCall(item) => item.as_legacy_end_event().into_iter().collect(),
-            TurnItem::Reasoning(item) => item.as_legacy_events(show_raw_agent_reasoning),
-            TurnItem::ContextCompaction(item) => vec![item.as_legacy_event()],
         }
     }
 }
