+++ codex-rs/protocol/src/lib.rs
@@ -16,6 +16,7 @@ pub mod dynamic_tools;
 pub mod error;
 pub mod exec_output;
 pub mod items;
+mod legacy_events;
 pub mod mcp;
 pub mod mcp_approval_meta;
 pub mod memory_citation;
